refactor(core): move tokenizer to core; update imports (P16 step 2)
