+++ ml_playground/experiments/shakespeare/preparer.py
@@ -14,7 +14,7 @@ from ml_playground.data_pipeline.transforms.io import (
     snapshot_file_states,
     write_bin_and_meta,
 )
-from ml_playground.tokenizer import create_tokenizer
+from ml_playground.core.tokenizer import create_tokenizer
 from ml_playground.experiments.protocol import (
     Preparer as _PreparerProto,
     PrepareReport,
